Upgrade to Gradle 2.12
Remove IT for Gradle < 2.8, because of introduction of Groovy 2.4. It creates binary incompatibility
with Spock (used for Groovy 2.4) and Groovy 2.3 available in older Gradle releases.
Tried to migrate to Gradle TestKit, but this is not so mature at the moment, and is a pain to configure...

Correctly test Gradle version when trying to activate the plugin. It was failing since Gradle 2.10.